Create differential expression pvalues (dep) object

This function creates a dep object from a pvalues data frame.

create_dep(
  sampledata,
  contrast,
  groups = NULL,
  pvalues,
  expression = NULL,
  p_col_suffix = "pvalues",
  padj_col_suffix = "padj",
  fc_col_suffix = "logFC",
  padjust_method = "BH",
  multi_group_prefix = NULL
)

Arguments

sampledata
A data frame containing the sample information. This must contain: an ID column containing the sample IDs which can be matched to the expression data and a contrast column containing the three-level factor used for contrasts.

contrast
The column name in `sampledata` which contains the three-level factor used for contrast.

groups
The groups to be compared (if NULL this defaults to levels(sampledata[, 'contrasts'])).

pvalues
A data frame containing three `p_col_suffix` and three `fc_col_suffix` columns: one for each comparison between groups. Similarly it can also contain: three optional `padj_col_suffix` columns (if NULL adjusted p values are calculated using `padjust_method``); and the 'p', 'padj and 'fc' columns for a three-way test, such as ANOVA or likelihood ratio test, defined by `multi_group_prefix`.

expression
Optional data frame containing expression data for downstream analysis and visualisation. The rows must contain probes which match the rows in pvalues and the columns must contain samples which match sampledata$ID.

p_col_suffix
The suffix word to define columns containing p values (default = 'pvalues').

padj_col_suffix
The suffix word to define columns containing adjusted p values (default = 'padj'). If NULL these will be calculated using padjust_method.

fc_col_suffix
The suffix word to define columns containing adjusted p-values (default = 'logFC').

padjust_method
The method used to calculate adjusted p values if padj_col_suffix is NULL (default = 'BH'). See p.adjust.

multi_group_prefix
Optional column prefix for statistics (p, padj, and fold change) across all three groups (typically ANOVA or likelihood ratio tests). default = NULL.

Value

Returns an S4 dep object containing:

'pvalues' A data frame containing the p-values, adjusted p-values, and log2(fold changes) for all three comparisons between groups in the contrast factor, as well as optional multi-group tests.

'sampledata' Sample data with column ID and contrast

'contrast' The column name in `sampledata`` which contains the three-group contrast factor used for comparisons.

'multi_group_test' Column name prefix for statistical tests between all three groups

'expression' An optional data frame or matrix containing the expression data

References

Lewis, Myles J., et al. (2019). Molecular portraits of early rheumatoid arthritis identify clinical and treatment response phenotypes. Cell reports, 28:9

Examples

library(volcano3Ddata)
data(syn_data)
syn_p_obj <- create_dep(sampledata = syn_metadata,
                    contrast = "Pathotype",
                    pvalues = syn_pvalues,
                    p_col_suffix ="pvalue",
                    fc_col_suffix = "log2FoldChange",
                    multi_group_prefix = "LRT",
                    expression = syn_rld)