An S4 class to define the polar coordinates and pvalues for polar differential expression plots

An S4 class to define the polar coordinates and pvalues for polar differential expression plots

Slots

sampledata

Sample data with ID and contrast column.

contrast

The column name in `sampledata`` which contains the three-group contrast factor used for comparisons.

pvalues

A data frame containing the p-values, and adjusted p-values, for all three comparisons between groups in the contrast factor, as well as optional fold changes and multi-group tests.

multi_group_test

Column name prefix for statistical tests between all three groups

expression

A data frame or matrix containing the expression data

polar

A data frame containing:

The axis score or mean expression for each of the three groups in comparison

The z-score polar coordinates: 'y_zscore', 'x_zscore' and 'r_zscore'

The fold-change polar coordinates: 'y_fc', 'x_fc' and 'r_fc'

'angle': The angle in radians for polar coordinates

'angle_degrees': The angle in degrees

'maxExp': The group with the highest expression

'sig': The significance category

non_sig_name

The category name for variables which are not significant