Extract a subset population

Subsets data according to the significance groups.

Usage

significance_subset(polar, significance = NULL, output = "pvals")

Arguments

polar

A polar object including expression data from groups of interest. Created by polar_coords.

significance

Which significance factors to subset to. If NULL all levels except 'ns' (non-significant) are selected.

output

What object to return. Options are "pvals", "padj", "data", "df" for subset dataframes, or "polar" to subset the entire 'volc3d' class object.

References

Lewis, Myles J., et al. (2019). Molecular portraits of early rheumatoid arthritis identify clinical and treatment response phenotypes. Cell reports, 28:9

Examples

data(example_data)
syn_polar <- polar_coords(outcome = syn_example_meta$Pathotype,
                          data = t(syn_example_rld))

subset <- significance_subset(syn_polar, "L+", "df")