Synovial differential expression of genes across pathotypes in PEAC cohort

A dataset containing the differential expression parameters between different pathotype groups for 81 synovial biopsies from the PEAC cohort.

syn_example_p

Format

A data frame with 100 rows representing the most significant genes/probes and 13 columns for each statistical parameter:

Gene

The gene name

Fibroid_Lymphoid_pvalue

pvalue from fibroid vs lymphoid comparison

Fibroid_Lymphoid_padj

adjusted pvalue from fibroid vs lymphoid comparison

Fibroid_Lymphoid_log2FoldChange

logarithmic fold change from fibroid vs lymphoid comparison

Lymphoid_Myeloid_pvalue

pvalue from lymphoid vs myeloid comparison

Lymphoid_Myeloid_padj

adjusted pvalue from lymphoid vs myeloid comparison

Lymphoid_Myeloid_log2FoldChange

logarithmic fold change from lymphoid vs myeloid comparison

Myeloid_Fibroid_pvalue

pvalue from myeloid vs fibroid comparison

Myeloid_Fibroid_padj

adjusted pvalue from myeloid vs fibroid comparison

Myeloid_Fibroid_log2FoldChange

logarithmic fold change from myeloid vs fibroid comparison

LRT_pvalue

pvalue from three-way likelihood ratio comparison

LRT_padj

adjusted pvalue from three-way likelihood ratio comparison

Source

https://www.cell.com/cell-reports/fulltext/S2211-1247(19)31007-1