Coordinates for Three Way Polar Plot

This function creates a data frame for downstream polar plots containing the p-values from a three-way group comparison.

polar_coords(
  dep,
  primary_colours = c("green3", "blue", "red"),
  secondary_colours = c("cyan", "purple", "gold2"),
  non_sig_colour = "grey60",
  non_sig_name = "Not Significant",
  significance_cutoff = 0.01
)

Arguments

dep
A dep object with the pvalues between groups of interest. Created by create_dep.

primary_colours
The colours for the primary groups. I.e. points up in only one of the three groups (default = c('blue', 'red', 'green3')).

secondary_colours
The colours for points in secondary groups. I.e. those up in two of the three groups (default = c("violet", "gold3", "cyan")).

non_sig_colour
Colour of non-significant points

non_sig_name
Name to assign non-significant points

significance_cutoff
Value defining the significance cut-off (pvalues below this will be coloured non_sig_colour)

Value

Returns an S4 polar object containing:

'polar' A data.frame containing:

The mean expression for each of the three groups in comparison

The z-score polar coordinates: 'y_zscore', 'x_zscore' and 'r_zscore'

The fold-change polar coordinates: 'y_fc', 'x_fc' and 'r_fc'

'angle': The angle in radians for polar coordinates

'angle_degrees': The angle in degrees

'hue': The continuous colour for volcano3D

'col': The discrete colour for volcano3D

'maxExp': The maximally expressed group

'sig': The significance group

'pvalues' A data frame containing the p-values, adjusted p-values, and log2(fold changes) for all three comparisons between groups in the contrast factor, as well as optional multi-group tests.

'sampledata' Sample data with column ID and contrast

'contrast' The column name in `sampledata`` which contains the three-group contrast factor used for comparisons.

'multi_group_test' Column name prefix for statistical tests between all three groups

'expression' An optional data frame or matrix containing the expression data

'non_sig_name' The category name for variables which are classed as not significant

References

Lewis, Myles J., et al. (2019). Molecular portraits of early rheumatoid arthritis identify clinical and treatment response phenotypes. Cell reports, 28:9

Examples

data(syn_data)
syn_p_obj <- create_dep(sampledata = syn_metadata,
                    contrast = "Pathotype",
                    pvalues = syn_pvalues,
                    p_col_suffix = "pvalue",
                    fc_col_suffix = "log2FoldChange",
                    multi_group_prefix = "LRT",
                    expression = syn_rld)
syn_polar <- polar_coords(dep = syn_p_obj)
table(syn_polar@polar$sig)

#> 
#>           Fibroid Fibroid+Lymphoid+  Fibroid+Myeloid+          Lymphoid 
#>               885                19              1504              1793 
#> Lymphoid+Myeloid+           Myeloid   Not Significant 
#>               500               119             11415