Plots grid objects for inspection using plotly

This function creates an interactive grids in polar and cylindrical coordinates

show_grid(grid, plot_height = 700, axis_angle = 0)

Arguments

grid
A grid object produced by polar_grid.

plot_height
The plot height in px (default=700),

axis_angle
The angle in radians at which to add axis (default=0).

Value

Returns a list containing a polar and cylindrical coordinate system.

References

Lewis, Myles J., et al. (2019). Molecular portraits of early rheumatoid arthritis identify clinical and treatment response phenotypes. Cell reports, 28:9

Examples

data(example_data)
syn_polar <- polar_coords(sampledata = syn_example_meta,
                       contrast = "Pathotype",
                       groups = NULL,
                       pvalues = syn_example_p,
                       expression = syn_example_rld,
                       p_col_suffix = "pvalue",
                       padj_col_suffix = "padj",
                       non_sig_name = "Not Significant",
                       multi_group_prefix = "LRT",
                       significance_cutoff = 0.01,
                       fc_col_suffix='log2FoldChange',
                       fc_cutoff = 0.3)

grid <- polar_grid(r_vector=syn_polar@polar$r_zscore,
        z_vector=-log(syn_polar@pvalues$LRT_pvalue),
        r_axis_ticks = NULL,
        z_axis_ticks = NULL)
p <- show_grid(grid)
p$polar
p$cyl