Plots grid objects for inspection using plotly

This function creates an interactive grids in polar and cylindrical coordinates

Usage

show_grid(grid, plot_height = 700, axis_angle = 0, z_axis_title_offset = 1.2)

Arguments

grid

A grid object produced by polar_grid.

plot_height

The plot height in px (default=700),

axis_angle

The angle in radians at which to add axis (default=0).

z_axis_title_offset

Offset for z axis title (default=1.2).

Value

Returns a list containing a polar and cylindrical coordinate system.

References

Lewis, Myles J., et al. (2019). Molecular portraits of early rheumatoid arthritis identify clinical and treatment response phenotypes. Cell reports, 28:9

Examples

data(example_data)
syn_polar <- polar_coords(outcome = syn_example_meta$Pathotype,
                          data = t(syn_example_rld))

grid <- polar_grid(r_vector=syn_polar@df[[1]]$r,
        z_vector=syn_polar@df[[1]]$z,
        r_axis_ticks = NULL,
        z_axis_ticks = NULL)
p <- show_grid(grid)
p$polar

p$cyl